requirements.txt
pexpect==4.9.0
pytest==8.3.5
pytest-mock==3.12.0 
PyYAML==6.0.2